Add benchmarks
