+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27, pypy, py35, py36
+envlist = py27, pypy, py35, py36, py27-benchmark, py36-benchmark
 
 [testenv]
 deps =
@@ -7,3 +7,17 @@ deps =
 commands =
     python setup.py --quiet clean develop
     pytest {posargs} --doctest-modules src/pghstore/_native.py --doctest-modules src/pghstore/__init__.py tests
+
+[testenv:py27-benchmark]
+deps =
+    pytest
+commands =
+    python setup.py --quiet clean develop
+    pytest -s {posargs} benchmark
+
+[testenv:py36-benchmark]
+deps =
+    pytest
+commands =
+    python setup.py --quiet clean develop
+    pytest -s {posargs} benchmark
